requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.4
django-allauth==0.61.1
django-appconf==1.0.6
django-bootstrap3==24.2
django-compat==1.0.15
django-debug-toolbar==4.3.0
django-js-asset==2.2.0
djangorestframework==3.15.1
idna==3.6
install==1.3.5
oauthlib==3.2.2
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
rcssmin==1.1.1
requests==2.31.0
requests-oauthlib==2.0.0
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
urllib3==2.2.1
